Extract a common Uri to the new utility library
