+++ MULTIPLATFORM.md
@@ -63,6 +63,7 @@ Compose Multiplatform core libraries can be published to local Maven with the fo
 1. Use these gradle properties to set the published libraries versions
 
 `-Pjetbrains.publication.version.CORE_BUNDLE`,
+`-Pjetbrains.publication.version.CORE_URI`,
 `-Pjetbrains.publication.version.COMPOSE`,
 `-Pjetbrains.publication.version.COMPOSE_MATERIAL3_ADAPTIVE`,
 `-Pjetbrains.publication.version.LIFECYCLE`,
@@ -73,7 +74,7 @@ Compose Multiplatform core libraries can be published to local Maven with the fo
 The default value for the version is `0.0.0-SNAPSHOT`
 
 And library groups:
-`-Pjetbrains.publication.libraries=CORE_BUNDLE,COMPOSE,COMPOSE_MATERIAL3_ADAPTIVE,LIFECYCLE,NAVIGATION,SAVEDSTATE,WINDOW`
+`-Pjetbrains.publication.libraries=CORE_BUNDLE,CORE_URI,COMPOSE,COMPOSE_MATERIAL3_ADAPTIVE,LIFECYCLE,NAVIGATION,SAVEDSTATE,WINDOW`
 
 The default value includes all libraries.
