+++ core/core-uri/build.gradle
@@ -1,0 +1,102 @@
+/*
+ * Copyright (C) 2024 The Android Open Source Project
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+import androidx.build.PlatformIdentifier
+import androidx.build.Publish
+import org.jetbrains.kotlin.gradle.plugin.KotlinPlatformType
+
+plugins {
+    id("AndroidXPlugin")
+    id("com.android.library")
+}
+
+androidXMultiplatform {
+    android()
+    desktop()
+    mac()
+    linux()
+    ios()
+    js { browser() }
+    wasm { browser() }
+
+    defaultPlatform(PlatformIdentifier.ANDROID)
+
+
+    sourceSets {
+        commonMain {
+            dependencies {
+                api(libs.kotlinStdlib)
+            }
+        }
+
+        jvmMain.dependsOn(commonMain)
+        jvmTest.dependsOn(commonTest)
+        jbMain.dependsOn(commonMain)
+        jbTest.dependsOn(commonTest)
+
+        desktopMain.dependsOn(jvmMain)
+        desktopMain.dependsOn(jbMain)
+        desktopTest.dependsOn(jvmTest)
+        desktopTest.dependsOn(jbTest)
+
+        androidMain.dependsOn(jvmMain)
+        androidTest.dependsOn(jvmTest)
+
+        nonJvmMain.dependsOn(jbMain)
+        nonJvmTest.dependsOn(jbTest)
+
+        nativeMain.dependsOn(nonJvmMain)
+        nativeTest.dependsOn(nonJvmTest)
+
+        webMain.dependsOn(nonJvmMain)
+        webTest.dependsOn(nonJvmTest)
+
+        targets.all { target ->
+            if (target.platformType == KotlinPlatformType.native) {
+                target.compilations["main"].defaultSourceSet.dependsOn(nativeMain)
+                target.compilations["test"].defaultSourceSet.dependsOn(nativeTest)
+            } else if (target.platformType in [
+                    KotlinPlatformType.js,
+                    KotlinPlatformType.wasm
+            ]) {
+                target.compilations["main"].defaultSourceSet.dependsOn(webMain)
+                target.compilations["test"].defaultSourceSet.dependsOn(webTest)
+            }
+        }
+    }
+}
+
+kotlin {
+    watchosArm64()
+    watchosArm32()
+    watchosX64()
+    watchosSimulatorArm64()
+    tvosArm64()
+    tvosX64()
+    tvosSimulatorArm64()
+    mingwX64()
+}
+
+android {
+    namespace "androidx.core.uri"
+}
+
+androidx {
+    name = "androidx.core:core-uri"
+    publish = Publish.SNAPSHOT_AND_RELEASE
+    inceptionYear = "2024"
+    description = "Provides Uri in Kotlin Multiplatform projects"
+}
