+++ core/core-uri/src/jbMain/kotlin/androidx/core/uri/UriCodec.kt
@@ -14,13 +14,11 @@
  * limitations under the License.
  */
 
-package androidx.navigation.internal
+package androidx.core.uri
 
 private class URISyntaxException(
     input: String, reason: String, index: Int
-) : Exception(
-    "$reason at index $index: $input"
-)
+) : Exception("$reason at index $index: $input")
 
 /**
  * Decodes “application/x-www-form-urlencoded” content.
@@ -73,10 +71,11 @@ internal object UriCodec {
      * invalid inputs. Else, U+FFFd is emitted to the output in place of invalid input octets.
      */
     fun decode(
-        s: String,
+        s: String?,
         convertPlus: Boolean = false,
         throwOnFailure: Boolean = false
-    ): String {
+    ): String? {
+        s ?: return null
         val builder = StringBuilder(s.length)
         appendDecoded(builder, s, convertPlus, throwOnFailure)
         return builder.toString()
