+++ libraryversions.toml
@@ -37,6 +37,7 @@ CORE_ANIMATION = "1.0.0-rc01"
 CORE_ANIMATION_TESTING = "1.0.0-rc01"
 CORE_APPDIGEST = "1.0.0-alpha01"
 CORE_BUNDLE = "1.0.0-alpha01"
+CORE_URI = "1.0.0-alpha01"
 CORE_GOOGLE_SHORTCUTS = "1.2.0-alpha02"
 CORE_HAPTICS = "1.0.0-alpha01"
 CORE_I18N = "1.0.0-alpha02"
@@ -302,6 +303,7 @@ ANNOTATION_JETBRAINS = { group = "org.jetbrains.compose.annotation-internal", at
 COLLECTION_JETBRAINS = { group = "org.jetbrains.compose.collection-internal", atomicGroupVersion = "versions.COMPOSE", overrideInclude = [ ":collection:collection" ] }
 
 CORE_BUNDLE_JETBRAINS = { group = "org.jetbrains.androidx.core", atomicGroupVersion = "versions.CORE_BUNDLE", overrideInclude = [ ":core:core-bundle" ] }
+CORE_URI_JETBRAINS = { group = "org.jetbrains.androidx.core", atomicGroupVersion = "versions.CORE_URI", overrideInclude = [ ":core:core-uri" ] }
 SAVEDSTATE_JETBRAINS = { group = "org.jetbrains.androidx.savedstate", atomicGroupVersion = "versions.SAVEDSTATE", overrideInclude = [ ":savedstate:savedstate" ] }
 WINDOW_CORE_JETBRAINS = { group = "org.jetbrains.androidx.window", atomicGroupVersion = "versions.WINDOW", overrideInclude = [ ":window:window-core" ]  }
