+++ navigation/navigation-common/build.gradle
@@ -56,6 +56,7 @@ androidXMultiplatform {
                 api project(":annotation:annotation")
                 api project(":collection:collection")
                 api project(":core:core-bundle")
+                api project(":core:core-uri")
                 api project(":lifecycle:lifecycle-common")
                 api project(":lifecycle:lifecycle-runtime")
                 api project(":lifecycle:lifecycle-viewmodel")
