+++ navigation/navigation-common/src/commonTest/kotlin/androidx/navigation/NavDeepLinkActionTest.kt
@@ -16,6 +16,7 @@
 
 package androidx.navigation
 
+import androidx.core.uri.UriUtils
 import androidx.kruth.assertThat
 import androidx.kruth.assertWithMessage
 import kotlin.test.Test
