+++ settings.gradle
@@ -232,6 +232,7 @@ includeProject(":lint-checks")
 includeProject(":lint-checks:integration-tests")
 
 includeProject(":core:core-bundle")
+includeProject(":core:core-uri")
 includeProject(":lifecycle:lifecycle-common")
 includeProject(":lifecycle:lifecycle-runtime")
 includeProject(":lifecycle:lifecycle-runtime-compose")
